# services/cart-service/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cart Service API",
    "version": "0.3.0",
    "description": "Serviço de carrinho de compras com Prisma + Postgres. Carrinhos por sessão; criar/obter carrinho, adicionar, alterar quantidade e remover itens."
  },
  "servers": [{ "url": "http://localhost:3003", "description": "Local dev" }],
  "tags": [
    { "name": "Health", "description": "Verificação de saúde do serviço" },
    { "name": "Cart", "description": "Operações sobre o carrinho de compras" }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "tags": ["Health"],
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Serviço operacional",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "service": { "type": "string", "example": "cart-service" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/carts/{sessionId}": {
      "get": {
        "summary": "Obter ou criar carrinho",
        "description": "Retorna o carrinho da sessão. Se não existir, cria um carrinho vazio e retorna.",
        "tags": ["Cart"],
        "operationId": "getOrCreateCart",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Carrinho (com items vazio ou preenchido)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Cart" }
              }
            }
          },
          "400": {
            "description": "sessionId ausente ou vazio",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/carts/{sessionId}/items": {
      "post": {
        "summary": "Adicionar item ao carrinho",
        "description": "Adiciona ou atualiza a quantidade do produto no carrinho (upsert por productId). Cria o carrinho se não existir.",
        "tags": ["Cart"],
        "operationId": "addItem",
        "parameters": [
          { "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["productId", "name", "price"],
                "properties": {
                  "productId": { "type": "integer", "example": 1 },
                  "name": { "type": "string", "example": "iPhone 15 Pro" },
                  "price": { "type": "number", "example": 219.9 },
                  "quantity": { "type": "integer", "default": 1, "minimum": 1, "example": 2 },
                  "imageUrl": { "type": "string", "format": "uri", "description": "URL da imagem do produto" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Carrinho atualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" } } }
          },
          "400": {
            "description": "Corpo inválido",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/carts/{sessionId}/items/{productId}": {
      "patch": {
        "summary": "Alterar quantidade do item",
        "description": "Atualiza a quantidade do produto no carrinho. Use quantity 0 para efeito de remoção (ou use DELETE).",
        "tags": ["Cart"],
        "operationId": "updateItemQuantity",
        "parameters": [
          { "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "quantity": { "type": "integer", "minimum": 0, "example": 3 } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Carrinho atualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" } } }
          },
          "400": {
            "description": "Parâmetros inválidos",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Carrinho ou item não encontrado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Remover item do carrinho",
        "tags": ["Cart"],
        "operationId": "removeItem",
        "parameters": [
          { "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Carrinho após remoção",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" } } }
          },
          "400": {
            "description": "Parâmetros inválidos",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Carrinho ou item não encontrado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cart": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "sessionId": { "type": "string" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CartItem" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "cartId": { "type": "integer" },
          "productId": { "type": "integer" },
          "name": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri", "nullable": true },
          "price": { "type": "string", "format": "decimal" },
          "quantity": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  }
}